+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
             android:name=".ui.addexpense.AddExpenseActivity"
             android:exported="true" />
         <activity
-            android:name=".ui.expensedetails.ExpenseDetailActivity"
+            android:name=".ui.expensedetail.ExpenseDetailActivity"
             android:exported="true" />
     </application>
