+++ app/src/main/java/com/kappzzang/jeongsan/ui/expensedetail/ExpenseDetailItemListAdapter.kt
@@ -1,4 +1,4 @@
-package com.kappzzang.jeongsan.ui.expensedetails
+package com.kappzzang.jeongsan.ui.expensedetail
 
 import android.content.Context
 import android.text.Editable
